Be more cautious about warnings
Only silent a few of them during testing
Activate -Werror during testing

As a result fix a few issues :-)

+++ pythran/pythonic/include/numpy/arange.hpp
@@ -23,7 +23,7 @@ namespace numpy
             index_{static_cast<long>(n / vector_type::size)}
       {
         T from[vector_type::size];
-        for (long i = 0; i < vector_type::size; ++i)
+        for (size_t i = 0; i < vector_type::size; ++i)
           from[i] = start + i * step;
         curr_ = xsimd::load_unaligned(from);
       }
